Front: Expanded Player: Rework on text
